% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_A.R
\docType{data}
\name{A}
\alias{A}
\title{Adjacency matrices of binary brain networks for 212 subjects}
\format{An object of class \code{array} of dimension 68 x 68 x 212.}
\source{
\href{https://github.com/wangronglu/CISE-algorithm}{Github}
}
\usage{
data(A)
}
\description{
Data from Human Connectome Project (HCP). The brain is segmented into 68
regions. Each subject's binary network consists of a collection of inter-
connections among those brain regions, with ones indicate a connection between a
pair of regions and zeros indicate no connection.
}
\references{
Wang et al. (2017)
  (\href{https://arxiv.org/abs/1707.06360}{Arxiv})
}
\keyword{datasets}
