% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_VSPLOT.R
\docType{data}
\name{VSPLOT}
\alias{VSPLOT}
\title{Visuospatial processing ability for 212 subjects}
\format{An object of class \code{character} of length 212.}
\source{
\href{https://github.com/wangronglu/CISE-algorithm}{Github}
}
\usage{
data(VSPLOT)
}
\description{
Data from Human Connectome Project (HCP). Visuospatial processing is commonly
assessed using the Variable Short Penn Line Orientation Test (VSPLOT), where
two line segments are presented on the screen and participants are asked to
rotate a movable line so that it is parallel to the fixed line. The data
contains an indicator of high/low visuospatial processing score for 212
subject. The order of the subjects is in the same order as in the array of
adjacency matrices data(A).
}
\references{
Wang et al. (2017)
  (\href{https://arxiv.org/abs/1707.06360}{Arxiv})
}
\keyword{datasets}
